homembit-iot/mqtt_IoT_Surfboard/src/requirements.txt
colored==1.4.2
paho-mqtt>=1.0.0
